user-input-agent/pyproject.toml
[project]
name = "user-input-agent"
version = "0.1.0"
description = "A user input analysis agent that processes and validates user-reported incidents from text reports, images, voice messages, and routes them to appropriate specialized agents using Cloud Functions and Natural Language AI."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0,<2.0.0",
    "google-cloud-aiplatform[adk,agent_engine]>=1.100.0",
    "google-cloud-language>=2.0.0",
    "google-cloud-speech>=2.0.0",
    "google-cloud-vision>=3.0.0",
    "google-cloud-functions>=1.0.0",
    "google-cloud-storage>=2.0.0",
    "python-dotenv>=1.0.0",
    "pathlib",
    "asyncio",
    "pillow>=8.0.0",
    "pydub>=0.25.0",
    "langdetect>=1.0.9",
    "textblob>=0.17.0",
    "spacy>=3.4.0",
    "regex>=2022.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false 